Lengthy explanation. Line by line.
Using the modular operator is included in this practice question. 
Line number 6 is the float which dignifies that this will be a decimal or a double.
Line number 7 we are setting 'a' and i as integers.
Line number 8 we are now setting 'a' to a new value of 4. When I say a new value, meaning 'a' could have been set to something else.
Line number 9 we are taking our value of 'a' and subtracting it by 1. We now have a new value for 'a' being 3.
Line number 10 is our first output statement. %d is the specifier for an integer value ( int variable) so we must print out an integer.
The print statement wants the integer value of 'a' to be printed out along with a space before the value of 'a'.
OUTPUT: P= 3
Line number 11 we are changing the value of 'a' to 5. NOTE a += 2; is equal to a = a+2; just a shorter version. 
Line number 12 we are increasing the value of 'a' by 1. Now the value of 'a' is now 6.  
Line number 13 is where the modular operator comes into play. This operator looks for the remainder. So the value for this is 0 as there is no remainder when dividing 6 by 2 so 'a' is now equal to 0. 
Line 14 is the second output statement with a spacing of 4 spaces and then the integer 0. This may be confusing but look closely and you will see %4d where the d is calling an integer as discussed before. 
OUTPUT:    0
Line 15 and 16 are introducing new